checkpoint, only the post processing after the Yes No dialog remaining

+++ dynamo/viewers/viewers_data.py
@@ -25,20 +25,14 @@
 # *
 # **************************************************************************
 
-from os.path import join, abspath
-import numpy as np
-
-from dynamo.utils import getCurrentTomoCountFile
+from os.path import join, abspath, exists
 from dynamo.viewers.DynamoTomoProvider import DynamoTomogramProvider
 from pwem.viewers import ObjectView
 import pyworkflow.viewer as pwviewer
 from pyworkflow.gui.dialog import askYesNo
-from pyworkflow.utils import removeBaseExt
 from pyworkflow.utils.properties import Message
 from tomo.objects import SetOfCoordinates3D, SetOfMeshes
-import tomo.constants as const
 from dynamo.viewers.views_tkinter_tree import DynamoTomoDialog
-from dynamo.convert import textFile2Coords
 from dynamo import VLL_FILE
 
 
@@ -63,17 +57,24 @@ class DynamoDataViewer(pwviewer.Viewer):
         if isinstance(obj, SetOfCoordinates3D) or isinstance(obj, SetOfMeshes):
             outputCoords = obj
             precedents = outputCoords.getPrecedents()
-            tomoIdDict = {tomo.getTsId(): tomo for tomo in precedents}
+            tomoIdDict = {tomo.getTsId(): tomo.clone() for tomo in precedents}
             tomoList = list(tomoIdDict.values())
+            nCoordsDict = {}
+            for tomo in tomoList:
+                tomoId = tomo.getTsId()
+                nCoordsDict[tomoId] = len(list(outputCoords.iterItems(where='_tomoId=="%s"' % tomoId)))
+
             path = self.protocol._getExtraPath()
-            tomoProvider = DynamoTomogramProvider(tomoList, path, nParticles=len(outputCoords))
+            tomoProvider = DynamoTomogramProvider(tomoList, path, nParticlesDict=nCoordsDict)
             listTomosFile = join(path, VLL_FILE)
 
-            # Create list of tomos file (VLL file)
-            with open(listTomosFile, 'w') as tomoFid:
-                for tomo in tomoList:
-                    tomoPath = abspath(tomo.getFileName())
-                    tomoFid.write(tomoPath + '\n')
+            # Create list of tomos file (VLL file), only required if using the Dynamo viewer in a protocol from any
+            # other plugin
+            if not(exists(listTomosFile)):
+                with open(listTomosFile, 'w') as tomoFid:
+                    for tomo in tomoList:
+                        tomoPath = abspath(tomo.getFileName())
+                        tomoFid.write(tomoPath + '\n')
 
             self.dlg = DynamoTomoDialog(self._tkRoot, path,
                                         provider=tomoProvider,
